Nerf the ParticleTypeRegistry
